fix: do not anti-alias Classic GroupBox Border
This appearance is more similar to Windows' GroupBox elements.
This fix is accomplished using two Border elements. One is the aliased-edge Border as seen in Sample. The other contains the templated content, but uses a transparent brush for the border. The former's layout data has bindings to the latter's layout data.

Fixes #92